Fix renaming xep_0045b to xep_0045
